.gitignore += *.so.*
This covers .so files created with soversion, e.g. there can be
libdemo.so and libdemo.so.1.0 . The latter is covered by added patern.

No need to change dylib and dll patterns - according to current
setuptools_dso code, the soversion is either used before '.dylib'
suffix, or, for '.dll' case is not present in file name at all.