Create info module for each DSO
Generates an info module per DSO built with information on the original
dot separated name, the stem and full path to the DSO. The info module
also adds the directory containing the DSO to the DLL path on Windows
once the info module is loaded.

The name for the info module is configurable per DSO and defaults to
`"{stem}_dso.py"`. Setting the name to `None` turns off info moduel
generation for the DSO.

The change does not attempt to create proper package structure to ensure
the info modules are importable, that is up to the user.